+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
-## [0.3.4] (in development)
+## [0.3.5] (in development)
+
+## [0.3.4] (2025-16-04)
+
+Thanks go to @wetneb, @anwaralameddin, @rudolphfroger, @jcamiel, @imcsk8 for contributions to this release.
 
 ### Added
 
@@ -11,6 +15,7 @@
 
 ### Changed
 
+* Call bindgen at build time on Unix platforms (thanks @wetneb)
 * Update the implementation of `StructuredError` so that all validation errors are returned from the validation methods present on `SchemaValidationContext`. Previously, all returned validation errors were identical due to libxml reusing a global memory address for all reported errors. Thanks @JDSeiler !
 * The `message` method of `StructuredError` has been deprecated.